Structure cpp extension
