+++ torch_ttnn/cpp_extension/core/TtnnCustomAllocator.hpp
@@ -1,9 +1,9 @@
+#pragma once
+
 #include <c10/core/Allocator.h>
 
-#include "extension_utils.hpp"
-
 // A dummy allocator used for creating Storage for Torch Tensors
-// This should not use any actual space
+// This should not use any actual space (data is stored in DRAM, not CPU RAM)
 struct TtnnCustomAllocator final : at::Allocator {
     TtnnCustomAllocator() = default;
     at::DataPtr allocate(size_t nbytes) const override;
@@ -12,3 +12,5 @@ struct TtnnCustomAllocator final : at::Allocator {
 
     at::DeleterFnPtr raw_deleter() const override;
 };
+
+TtnnCustomAllocator& get_ttnn_custom_allocator();
