+++ torch_ttnn/cpp_extension/core/TtnnGuard.hpp
@@ -1,8 +1,10 @@
 #pragma once
 
 #include <c10/core/DeviceGuard.h>
-#include "extension_utils.hpp"
-#include "ttnn/device.hpp"
+
+#include <ttnn/device.hpp>
+
+#include "utils/extension_utils.hpp"
 
 // =====================================
 // ============= Device Guards =========
