+++ torch_ttnn/cpp_extension/core/TtnnTensorImpl.cpp
@@ -1,7 +1,9 @@
-#include "TtnnTensorImpl.hpp"
-#include "ttnn/device.hpp"
-#include "ttnn/operations/creation.hpp"
-#include "TtnnGuard.hpp"
+#include <ttnn/device.hpp>
+#include <ttnn/operations/creation.hpp>
+
+#include "core/TtnnTensorImpl.hpp"
+
+#include "utils/extension_utils.hpp"
 
 namespace {
 std::vector<int64_t> get_row_major_strides(const std::vector<int64_t>& sizes) {
