+++ torch_ttnn/cpp_extension/ops/unary.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <ATen/core/Tensor.h>
+
+namespace tt_eager::ops::unary {
+at::Tensor& ttnn_abs_out(const at::Tensor& input, at::Tensor& out);
+}
