+++ torch_ttnn/cpp_extension/utils/extension_utils.hpp
@@ -1,12 +1,22 @@
 #pragma once
 
-#include <vector>
+#include <c10/core/ScalarType.h>
+#include <c10/util/Exception.h>
+#include <fmt/format.h>
+#include <ttnn/operations/core/core.hpp>
 
-#include "c10/core/ScalarType.h"
-#include "c10/util/Exception.h"
-#include "ATen/core/Tensor.h"
-
-#include "ttnn/operations/core/core.hpp"
+template <typename T>
+std::ostream& operator<<(std::ostream& os, const std::vector<T>& vec) {
+    os << "[";
+    for (size_t i = 0; i < vec.size(); ++i) {
+        os << vec[i];
+        if (i != vec.size() - 1) {
+            os << ", ";
+        }
+    }
+    os << "]";
+    return os;
+}
 
 template <typename Arg, typename... Args>
 void doPrint(
@@ -33,12 +43,6 @@ inline ttnn::DataType dtype_torch_to_ttnn(at::ScalarType dtype) {
         case at::ScalarType::BFloat16: return ttnn::DataType::BFLOAT16;
         default: return ttnn::DataType::BFLOAT16;
     }
-}
-
-inline void convert_vector_from_uint32_to_int(std::vector<int>& dst_vector, const std::vector<uint32_t>& src_vector) {
-    std::transform(src_vector.begin(), src_vector.end(), std::back_inserter(dst_vector), [](const uint32_t value) {
-        return static_cast<int>(value);
-    });
 }
 
 #define GET_OPTIONAL_OR_ASSERT(optional)             \
